+++ src/callbacks.hpp
@@ -11,13 +11,6 @@
 #pragma once
 
 #include <idp.hpp>
-#include <list>
-
-#include "globals.hpp"
 
 ssize_t idaapi tracer_callback(void * /*user_data*/, int notification_code, va_list va);
 ssize_t idaapi ui_callback(void * /*ud*/, int notification_code, va_list va);
-
-int tritonize(ea_t pc, thid_t threadID = 0);
-void set_SMT_results(Input* input_ptr);
-void triton_restart_engines();
