+++ src/globals.hpp
@@ -92,25 +92,6 @@ struct cmdOptionStruct{
 };
 extern struct cmdOptionStruct cmdOptions;
 
-class Input
-{
-public:
-	int bound;
-	std::vector <triton::arch::MemoryAccess> memOperand;
-	std::vector <triton::arch::Register> regOperand;
-
-	//! Constructor.
-	Input()
-	{
-		this->bound = 0;
-	}
-
-	//! Destructor.
-	~Input()
-	{
-	}
-};
-
 //This variable stores the ea where the popup menu is being displayed, this is use to add the submenus for Solve condition
 extern ea_t popup_menu_ea;
